+++ solutions/Continuous_model/deprecated/train_concurrent.py
@@ -15,8 +15,8 @@ from tensordict import TensorDict
 
 from torchrl.modules import MaskedCategorical
 
-from solutions.Td_petri.enviroment import Env_PN_Concurrent
-from solutions.PPO.network.models import DualHeadPolicyNet
+from solutions.Continuous_model.deprecated.env import Env_PN_Concurrent
+from solutions.model.network import DualHeadPolicyNet
 from data.ppo_configs.training_config import PPOTrainingConfig